backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "renderstudio-backend"
version = "0.1.0"
description = "RenderStudio FastAPI backend"
requires-python = ">=3.11"
readme = "README.md"
dependencies = [
  "fastapi>=0.110,<1.0",
  "uvicorn[standard]>=0.29",
  "pydantic>=2.6,<3.0",
  "pydantic-settings>=2.2",
  "SQLAlchemy[asyncio]>=2.0",
  "alembic>=1.13",
  "asyncpg>=0.29",
  "aiosqlite>=0.19",
  "redis>=5.0",
  "celery>=5.3",
  "minio>=7.2",
  "anthropic>=0.29",
  "structlog>=24.1",
  "sse-starlette>=2.1",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "authlib>=1.3",
  "python-jose[cryptography]>=3.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "ruff>=0.4",
  "mypy>=1.10",
  "types-redis",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
